move fixlist code out of x86 files to generic files
